Ficed critical bug with Text copy semantics
How the copy semantics for Text work is that, since it's a little difficult to extract data from a texture (AFAIK), the texture is regenerated from the copied data. Problem is, it'll fucking die if it's told to render *nothing*. So, as you can see from the diff, I wasn't checking to make sure that there was something to render before going and attempting to generate a texture. Now we check if there are more than 0 lines, which is garunteed to tell if there's something to render.